+++ spec/fixtures/status_publisher.rb
@@ -1,18 +1,21 @@
 # frozen_string_literal: true
 
+require 'low_type'
 require_relative '../../lib/observers'
 
-class ClassPublisher
+class StatusPublisher
+  include LowType
   include Observers
-  observable
+
+  observable Status[200]
 
   class << self
     def trigger_action(action:)
-      trigger(action:)
+      trigger(LowType::Status[200], action:)
     end
 
     def trigger_event(event:)
-      trigger(event:)
+      trigger(LowType::Status[200], event:)
     end
   end
 end