+++ spec/fixtures/status_subscriber.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require 'low_type'
+require_relative '../../lib/observers'
+
+class StatusSubscriber
+  include LowType
+  include Observers
+
+  observe Status[200]
+end